# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel API",
        "description": "Sayohat veb-sayti uchun API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://toqtarbay.dbc.uz/api",
            "description": "Production Server"
        },
        {
            "url": "http://travel.test/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Barcha sharhlarni olish",
                "description": "Aktiv sharhlarni ro'yxatini qaytaradi. Tour bo'yicha filter qilish mumkin.",
                "operationId": "1977cd84258463d927b22fbdb185d50b",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Til kodi (uz, ru, kk, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "uz",
                            "enum": [
                                "uz",
                                "ru",
                                "kk",
                                "en"
                            ]
                        }
                    },
                    {
                        "name": "tour_id",
                        "in": "query",
                        "description": "Tur ID (ixtiyoriy)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli javob",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_name": {
                                                        "type": "string",
                                                        "example": "Alisher Navoiy"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Toshkent"
                                                    },
                                                    "comment": {
                                                        "type": "string",
                                                        "example": "Ajoyib tur edi! Juda yoqdi, hamma narsasi mukammal tashkil etilgan."
                                                    },
                                                    "rating": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "video_url": {
                                                        "type": "string",
                                                        "example": "https://youtube.com/watch?v=xxx",
                                                        "nullable": true
                                                    },
                                                    "tour": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Orol dengizi safari"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/reviews/{id}": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Bitta sharhni olish",
                "description": "ID bo'yicha sharhni qaytaradi",
                "operationId": "acf25d40c9352f206d9be8f3285fe724",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Til kodi (uz, ru, kk, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "uz",
                            "enum": [
                                "uz",
                                "ru",
                                "kk",
                                "en"
                            ]
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sharh ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli javob",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_name": {
                                                    "type": "string",
                                                    "example": "Alisher Navoiy"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "Toshkent"
                                                },
                                                "comment": {
                                                    "type": "string",
                                                    "example": "Ajoyib tur edi! Juda yoqdi, hamma narsasi mukammal tashkil etilgan."
                                                },
                                                "rating": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "video_url": {
                                                    "type": "string",
                                                    "example": "https://youtube.com/watch?v=xxx",
                                                    "nullable": true
                                                },
                                                "tour": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "title": {
                                                            "type": "string",
                                                            "example": "Orol dengizi safari"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sharh topilmadi"
                    }
                }
            }
        },
        "/reviews/banner": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Reviews sahifasi bannerini olish",
                "operationId": "0427635b8566022c09983e196197c107",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Til kodi (kk, uz, ru, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "uz"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli javob",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Mijozlar sharhlari"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "/storage/uploads/review-banner.jpg"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Banner topilmadi"
                    }
                }
            }
        },
...
    },
...
}